% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tools.R
\name{combine_tools}
\alias{combine_tools}
\title{Combine data for quartile figure}
\usage{
combine_tools(
  KRSA_df = NULL,
  UKA_df = NULL,
  KEA3_df = NULL,
  PTM_SEA_df = NULL,
  mapping_df = kinome_mp_file_v1
)
}
\arguments{
\item{KRSA_df}{dataframe, KRSA table output (requires at least Kinase and Score columns)}

\item{UKA_df}{dataframe, UKA table output (requires at least Kinase and Score columns)}

\item{KEA3_df}{dataframe, KEA table output (requires at least Kinase and Score columns)}

\item{PTM_SEA_df}{dataframe, PTM_SEA table output (requires at least Kinase and Score columns)}

\item{mapping_df}{kinome mapping df (default is kinome_mp_file_v1)}
}
\value{
dataframe, ready for quartile figure
}
\description{
reads ranked tables from the different tools (KRSA, UKA, ... etc)
}
\details{
This function takes in ranked tables from the different tools (KRSA, UKA, ... etc) and map them to the kinome mapping file and return df ready for the quartile figure
}
