% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creedenzymatic.R
\name{creedenzymatic}
\alias{creedenzymatic}
\title{Runs Creedenzymatic}
\usage{
creedenzymatic(
  KRSA_table,
  UKA_table,
  LFC_table,
  avg_krsa = T,
  avg_lfc = T,
  prefix = "Comp1",
  ...
)
}
\arguments{
\item{KRSA_table}{dataframe, KRSA table output}

\item{UKA_table}{dataframe, UKA table output}

\item{LFC_table}{dataframe, KEA table output}

\item{..., }{arguments passed to other functions}
}
\value{
dataframe, Ranked and quartiled table
}
\description{
reads KRSA, UKA, LFC tables and run creedenzymatic
}
\details{
This function takes in table and rank and quartile kinases based on the absolute Score values
}
