% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_io.R
\name{.read.gctx.meta}
\alias{.read.gctx.meta}
\title{Parse row or column metadata from GCTX files}
\usage{
.read.gctx.meta(gctx_path, dimension = "row", ids = NULL)
}
\arguments{
\item{gctx_path}{the path to the GCTX file}

\item{dimension}{which metadata to read (row or column)}

\item{ids}{a character vector of a subset of row/column ids
for which to read the metadata}
}
\value{
a \code{data.frame} of metadata
}
\description{
Parse row or column metadata from GCTX files
}
\seealso{
Other GCTX parsing functions: 
\code{\link{.append.dim}()},
\code{\link{.fix.datatypes}()},
\code{\link{.parse.gctx}()},
\code{\link{.process_ids}()},
\code{\link{.read.gctx.ids}()},
\code{\link{.write.gctx.meta}()},
\code{\link{.write.gctx}()},
\code{\link{.write.gct}()}
}
\concept{GCTX parsing functions}
