% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_io.R
\name{.update.gctx}
\alias{.update.gctx}
\title{Update the matrix of an existing GCTX file}
\usage{
.update.gctx(x, ofile, rid = NULL, cid = NULL)
}
\arguments{
\item{x}{an array of data}

\item{ofile}{the filename of the GCTX to update}

\item{rid}{integer indices or character ids of the rows
to update}

\item{cid}{integer indices or character ids of the columns
to update}
}
\description{
Update the matrix of an existing GCTX file
}
\details{
Overwrite the rows and columns of \code{ofile}
as indicated by \code{rid} and \code{cid} respectively.
\code{rid} and \code{cid} can either be integer indices
or character ids corresponding to the row and column ids
in \code{ofile}.
}
