% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_io.R
\name{.write.grp}
\alias{.write.grp}
\title{Write a vector to a GRP file}
\usage{
.write.grp(vals, fname)
}
\arguments{
\item{vals}{the vector of values to be written}

\item{fname}{the desired file name}
}
\description{
Write a vector to a GRP file
}
\seealso{
\link{http://clue.io/help} for details on the GRP file format

Other CMap parsing functions: 
\code{\link{.parse.gmt}()},
\code{\link{.parse.gmx}()},
\code{\link{.parse.grp}()},
\code{\link{.write.gmt}()}
}
\concept{CMap parsing functions}
