% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_top_kinases.R
\name{extract_top_kinases}
\alias{extract_top_kinases}
\title{Extract Top Kinases}
\usage{
extract_top_kinases(combined_df, min_qrt, min_counts)
}
\arguments{
\item{combined_df}{dataframe, Ranked and quartiled dataframe}

\item{min_qrt}{integer, minimum quartile to count}

\item{min_counts}{integer, number of minimum hits}
}
\value{
vector, top kinases
}
\description{
reads combined dataframe (ranked and quartiled) and extracts top kinases based on adjustable criteria
}
\details{
This function takes in the combined dataframe (ranked and quartiled) and extracts top kinases based on adjustable criteria
}
