% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinome_mp_file}
\alias{kinome_mp_file}
\title{CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA)}
\format{
A data frame with 527 rows and 26 variables:
\describe{
}
}
\usage{
kinome_mp_file
}
\description{
A data frame of CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA) (Latest Version)
}
\keyword{datasets}
