% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinome_mp_file_v1}
\alias{kinome_mp_file_v1}
\title{CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA)}
\format{
A data frame with 503 rows and 14 variables:
\describe{
}
}
\usage{
kinome_mp_file_v1
}
\description{
A data frame of CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA) (Version 1)
}
\keyword{datasets}
