% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinome_mp_file_v2}
\alias{kinome_mp_file_v2}
\title{CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA)}
\format{
A data frame with 514 rows and 12 variables:
\describe{
}
}
\usage{
kinome_mp_file_v2
}
\description{
A data frame of CDRL Complete mapping file (UKA+KRSA+KEA3+PTM-SEA) (Version 2)
}
\keyword{datasets}
