% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptk_pamchip_86402_array_layout_ptmsea}
\alias{ptk_pamchip_86402_array_layout_ptmsea}
\title{CDRL Complete mapping of peptides - used for ptm-sea (PTK PamChip 86402)}
\format{
A data frame with x rows and x variables:
\describe{
}
}
\usage{
ptk_pamchip_86402_array_layout_ptmsea
}
\description{
A data frame of CDRL Complete mapping CDRL Complete mapping of peptides - used for ptm-sea (PTK PamChip 86402)
}
\keyword{datasets}
