% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartile_figure.R
\name{quartile_figure}
\alias{quartile_figure}
\title{Plot Quartile Figure}
\usage{
quartile_figure(df, df2)
}
\arguments{
\item{df2}{dataframe, KRSA ranked table}

\item{df1}{dataframe, UKA ranked table}
}
\value{
ggplot figure
}
\description{
Takes KRSA and UKA ranked tables and generate a quartile figure
}
\details{
This function takes KRSA and UKA ranked tables and generate a quartile figure
}
