% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_utils.R
\name{rank.gct}
\alias{rank.gct}
\title{Convert a GCT object's matrix to ranks}
\usage{
rank.gct(g, dim = "col", decreasing = T)
}
\arguments{
\item{g}{the \code{GCT} object to rank}

\item{dim}{the dimension along which to rank
(row or column)}

\item{decreasing}{boolean indicating whether
higher values should get lower ranks}
}
\value{
a modified version of \code{g}, with the
  values in the matrix converted to ranks
}
\description{
Convert a GCT object's matrix to ranks
}
\seealso{
Other GCT utilities: 
\code{\link{annotate.gct}()},
\code{\link{melt.gct}()},
\code{\link{merge.gct}()},
\code{\link{subset.gct}()}
}
\concept{GCT utilities}
