% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_kinases.R
\name{rank_kinases}
\alias{rank_kinases}
\title{Rank Kinases based on a score}
\usage{
rank_kinases(
  df,
  trns = c("raw", "abs"),
  sort = c("desc", "asc"),
  tool = c("KRSA", "UKA")
)
}
\arguments{
\item{df, }{dataframe with 2 columns: Kinase, Score}

\item{trns, }{for transformation of the score, the values accepted for this argument are abs and raw (abs: use absolute values of scores, raw: no transformation)}

\item{sort, }{accepts either asc or desc (ascending and descending)}

\item{tool, }{specifying the name of the tool}
}
\description{
This function will scale the scores on a percentile and quartile scales
}
