% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kea.R
\name{read_kea}
\alias{read_kea}
\title{Reads a dataframe of Peptides IDs and their Scores and run KEA3}
\usage{
read_kea(
  df,
  filter,
  cutoff,
  cutoff_abs = T,
  direction = "higher",
  lib = c("kinases"),
  ...
)
}
\arguments{
\item{df}{dataframe, must have at least Peptide and Score columns}

\item{filter}{boolean to subset peptides or not}

\item{cutoff}{numeric to act as the cutoff to filter out peptides}

\item{cutoff_abs}{boolean (use absolute value or not) default is TRUE}

\item{direction}{("lower", "higher) filter based on less than or bigger than the cutoff values (default to "higher")}

\item{lib}{searched kea libraries (default is "kinases" which will return only kinase libraries like ChengKSIN, PTMsigDB, PhosDAll)}

\item{..., }{arguments passed to rank_kinases function}
}
\value{
dataframe, Ranked and quartiled table
}
\description{
reads a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them
}
\details{
This function a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them
}
