% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ptmsea.R
\name{read_ptmsea}
\alias{read_ptmsea}
\title{Reads a dataframe of Peptides IDs and their Scores and run PTM-SEA}
\usage{
read_ptmsea(df, ...)
}
\arguments{
\item{df}{dataframe, must have at least Peptide and Score columns}

\item{..., }{arguments passed to run ptm-sea function}

\item{lib}{searched PTM-SEA libraries "kinase-substrate" or "all" (default is "kinase-substrate" which will return only kinase libraries like ChengKSIN, PTMsigDB, PhosDAll)}
}
\value{
dataframe, Ranked and quartiled table
}
\description{
reads a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run PTM-SEA
}
\details{
This function a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run PTM-SEA
}
