% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_kinases.R, R/read_uka.R
\name{read_uka}
\alias{read_uka}
\title{Maps kinases}
\usage{
read_uka(df)

read_uka(df)
}
\arguments{
\item{df}{dataframe, UKA table output (requires at least Kinase and Z columns)}
}
\value{
dataframe, mapped table

dataframe, Ranked and quartiled UKA table
}
\description{
reads KRSA or UKA table and map kinases it to uniprot

reads UKA table and checks for correct format
}
\details{
This function takes in UKA or KRSA table map kinases to uniprot names

This function takes in UKA table and rank and quartile kinases based on the absolute Score values
}
