% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kea.R
\name{run_kea}
\alias{run_kea}
\title{Run KEA3 API based on a set of gene symbols}
\usage{
run_kea(gene_set, lib = "kinase-substrate")
}
\arguments{
\item{gene_set}{vector, HGNC gene symbols based on the differentially phosphorylated peptides}

\item{lib}{searched kea libraries "kinase-substrate" or "all" (default is "kinase-substrate" which will return only kinase libraries like ChengKSIN, PTMsigDB, PhosDAll)}
}
\value{
list, tables from each KEA3 library
}
\description{
This function takes in HGNC gene symbols and connect to KEA3 API and returns results
}
