% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ptmsea.R
\name{run_ptmsea}
\alias{run_ptmsea}
\title{Run PTM-SEA API using a gct file as input}
\usage{
run_ptmsea(gct_object, lib = "iptmnet", nperm = 1000, min.overlap = 1, ...)
}
\arguments{
\item{lib}{searched kea libraries "iptmnet" or "ptm-sea" or "all (default is "iptmnet" which uses the iptmnet mapping)}

\item{nperm}{number of permutations}

\item{min.overlap}{minimum overlap of target peptides with referernce peptides sets}

\item{...}{additional arguments passed to the ssGSEA_ce function}

\item{gene_set}{vector, HGNC gene symbols based on the differentially phosphorylated peptides}
}
\value{
list
}
\description{
This function takes in a gct file (created by the read_prmsea function) and run PTM-SEA API and returns results
}
