% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stk_pamchip_87102_array_layout_ptmsea}
\alias{stk_pamchip_87102_array_layout_ptmsea}
\title{CDRL Complete mapping of peptides - used for ptm-sea (STK PamChip 87102)}
\format{
A data frame with x rows and x variables:
\describe{
}
}
\usage{
stk_pamchip_87102_array_layout_ptmsea
}
\description{
A data frame of CDRL Complete mapping CDRL Complete mapping of peptides - used for ptm-sea (STK PamChip 87102)
}
\keyword{datasets}
