% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uka_db_full}
\alias{uka_db_full}
\title{UKA Complete DB mapping File (STK + PTK)}
\format{
A data frame with 11385 rows and 17 variables:
\describe{
}
}
\usage{
uka_db_full
}
\description{
A data frame of the UKA Complete DB mapping File (STK + PTK)
}
\keyword{datasets}
