% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartile_figure.R
\name{quartile_figure}
\alias{quartile_figure}
\title{Plot quartile Figure}
\usage{
quartile_figure(df, grouping = "KinaseFamily")
}
\arguments{
\item{df}{dataframe, combined mapped tables}

\item{grouping}{character to choose grouping (KinaseFamily or group). Default is KinaseFamily}
}
\value{
ggplot figure
}
\description{
Takes the combined ranked dataframe (KRSA, UKA, .. etc) and generate a quartile figure
}
