% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_kinases.R
\name{rank_kinases}
\alias{rank_kinases}
\title{Rnak Kinases based on a score}
\usage{
rank_kinases(
  df,
  trns = c("raw", "abs"),
  sort = c("desc", "asc"),
  tool = c("KRSA", "UKA")
)
}
\arguments{
\item{df, }{dataframe}
}
\value{

}
\description{
Rnak Kinases based on a score
}
\examples{


}
