% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kea.R
\name{read_kea}
\alias{read_kea}
\title{Reads a dataframe of Peptides IDs and their Scores and run KEA3}
\usage{
read_kea(
  df,
  filter = T,
  cutoff = 0.2,
  cutoff_abs = T,
  direction = "higher",
  rm_duplicates = T,
  method = "MeanRank",
  lib = c("kinase-substrate"),
  ...
)
}
\arguments{
\item{df}{dataframe, must have at least Peptide and Score columns}

\item{filter}{boolean to subset peptides or not}

\item{cutoff}{numeric to act as the cutoff to filter out peptides}

\item{cutoff_abs}{boolean (use absolute value or not) default is TRUE}

\item{direction}{("lower", "higher) filter based on lower than or higher than the cutoff values (default to "higher")}

\item{rm_duplicates}{boolean (TRUE or FALSE) remove genes duplicates}

\item{method}{"MeanRank" takes the mean rank across all libraries or "MeanFDR" takes the mean of FDR across all libraries (default is "MeanRank")}

\item{lib}{searched kea libraries "kinase-substrate" or "all" (default is "kinase-substrate" which will return only kinase libraries like ChengKSIN, PTMsigDB, PhosDAll)}

\item{..., }{arguments passed to rank_kinases function}
}
\value{
dataframe, Ranked and quartiled table
}
\description{
reads a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them
}
\details{
This function a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them
}
