% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_krsa.R
\name{read_krsa}
\alias{read_krsa}
\title{Reads and Rank KRSA table}
\usage{
read_krsa(df)
}
\arguments{
\item{df}{dataframe, KRSA table output (requires at least Kinase and Z columns)}
}
\value{
dataframe, Ranked and quartiled KRSA table
}
\description{
reads KRSA table and checks for correct format
}
\details{
This function takes in KRSA table and rank and quartile kinases based on the absolute Score values
}
