% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_krsa.R
\name{read_krsa}
\alias{read_krsa}
\title{Read KRSA table}
\usage{
read_krsa()
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
`read_krsa` read krsa table and check for correct format
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
