% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uka.R
\name{read_uka}
\alias{read_uka}
\title{Reads and Rank UKA table}
\usage{
read_uka(df, ...)
}
\arguments{
\item{df}{dataframe, UKA table output (requires at least Kinase and Z columns)}

\item{..., }{arguments passed to rank_kinases function}
}
\value{
dataframe, Ranked and quartiled UKA table
}
\description{
reads UKA table and checks for correct format
}
\details{
This function takes in UKA table and rank and quartile kinases based on the absolute Score values
}
